Retrieve Database Data and Send to App
The previous version received requests from the app and returned a hardcoded string for login information and no database data was actually retrieved. Now the file has been edited so that the server retrieves login information from the database and sends it back to the app to process. The app then authenticates users based on this login information and then gives them access to the app's functionalities. (no register functionality yet)